Adding spell animations
1st Rank:
bane - does require an aura effect that's not default to Bane, unsure how to export item files otherwise would have included aura effect item. has allowance for free and patreon versions
bless - might be a bit load heavy
infuse-vitality - has an on-cast bolt toward targets and a persistent on the effect
spiritual armament - has a simple version for free, for patreon version has different animations for different damage types. The item:slug: gives a warning but thats the only way I could get it to work because of the selector for the roll type