`ArrayAnimationSet` —> `AnimationSetData[]`
Looking at the functions, the typing was just making some things messy for the very minor convenience of not adding `[]` at the end of the type. This has been updated so that `AnimationSetData` refers to *one particular* object (i.e. `AnimationSetData[] === ArrayAnimationSet`).

Also allows us to start using `Extract<>` on the new discriminated union format to narrow workflow types even harder.

Also renamed the `key` property to `rollOption` because, like, that's what it is and I was getting confused 😩